Fix life cycle issues in the listener
Fix life cycle issues in the RabbitMQ ChannelListener
